Create userToComputeEngineAPI.java
created the first API, this APIis An API between the user and the compute engine. The user will specify a source for the input, delimiter characters for the output, and a destination for the output.  This is currently not functional because source class and input class are not created yet