Feature (#8)
* Update testingSetup.java

* step 3 part 4

tester class should now be working. This shouldn't fail any  status checks and should work!!

* Update testingSetup.java

added copyright checkstyle feature

* Update testingSetup.java

Updated name of class and file name, updated indentation, deleted two lines

* Update testingSetup.java

Changed code on two lines hoping it works

* Rename testingSetup.class to TestingSetup.class

Name change

* Rename testingSetup.java to TestingSetup.java

Changed Name

* Update TestingSetup.java

updated TestingSetup (feature branch) trying to get this to main

* updated both javafile and setupfile

updated TestingSetup (feature branch) trying to get this to main

* Update TestingSetup.java

Fixed Header

* Update TestingSetup.java

Removed REGEX Stuff now that they gave been dropped

* Update README.md

updated README to explain what we are going to be computing for our project.

* Create userToComputeEngineAPI.java

created the first API, this APIis An API between the user and the compute engine. The user will specify a source for the input, delimiter characters for the output, and a destination for the output.  This is currently not functional because source class and input class are not created yet

* Update README.md

Added image for Compute Engine.

* added Input & source java classes

added Input & source java classes with no implimentation. added a .gitignore to remove all .class files (was requested by the professor). also changed the api's enum to be less intensive

* renamed UserToComputeEngineAPI to have camel case

renamed userToComputeEngineAPI  to UserToComputeEngineAPI , hopefully this works

* Delete TestingSetup.java

* Delete TestingSetup.class

---------

Co-authored-by: Henry <henrybecker1014@gmail.com>
Co-authored-by: Henry <125415992+Nerkled@users.noreply.github.com>